clean up warnings, add function to test if srid is geographic, and commit on execution of pre-block-sql

+++ apps/oci_wrapper.cpp
@@ -1660,7 +1660,7 @@ bool CheckError( sword nStatus, OCIError* hError )
         }
 
         CPLError( CE_Failure, CPLE_AppDefined, "%.*s",
-            sizeof(szMsg), szMsg );
+           (int) sizeof(szMsg), szMsg );
         break;
     
     default:
@@ -1676,7 +1676,7 @@ bool CheckError( sword nStatus, OCIError* hError )
                 (ub4) sizeof(szMsg), OCI_HTYPE_ERROR);
 
             CPLError( CE_Failure, CPLE_AppDefined, "%.*s",
-                sizeof(szMsg), szMsg );
+                (int) sizeof(szMsg), szMsg );
             break;
 
     }
